requirements.txt
asgiref==3.4.1
Django==3.2.23
sqlparse==0.4.4
psycopg[binary]==3.0.18
psycopg2==2.8.6
pillow==8.4.0
djangorestframework==3.12.4
django-cors-headers==3.8.0
djangorestframework-simplejwt==4.4.0
PyJWT==v1.7.1
stripe==8.5.0
requests_mock==1.9.3